pyproject.toml
[tool.poetry]
name = "oss-pulse"
version = "0.1.0"
description = ""
authors = ["Mebarek <mohammedmecheter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
pocketbase = "^0.12.1"
requests = "^2.32.3"
pandas = "^2.2.2"
plotly = "^5.24.0"
pygithub = "^2.4.0"
jupyter = "^1.1.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.5.1"
numpy = "^2.1.0"
python-dotenv = "^1.0.1"
ratelimit = "^2.2.1"
pytest = "^8.3.2"
apscheduler = "^3.10.4"
pyarrow = "^17.0.0"
streamlit-lottie = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
